Added fullPage.js, scrolling now works. MESSY HTML
Included fullPage scripts and stylesheets in functions.php.
Modified index.php to work with fullPage (added sections, div #page
is the wrapper), hard coded header to be outside fullpage wrapper
and hard coded footer to be inside the wrapper (last section).
Create a custom and clean template in the future!